+++ r-aci.tf
@@ -108,13 +108,12 @@ resource "azurerm_container_group" "aci" {
   }
 
   dynamic "identity" {
-    for_each = try(var.identity[*], [])
+    for_each = var.identity[*]
     content {
       type         = var.identity.type
       identity_ids = var.identity.identity_ids
     }
   }
-
 
   tags = merge(local.default_tags, var.extra_tags)
 }