+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Unreleased
 
 Added
-  * [GH-5](https://github.com/claranet/terraform-azurerm-aci/pull/5): Add Liveness and readiness probes
-  * [GH-5](https://github.com/claranet/terraform-azurerm-aci/pull/5): Add DNS Name Label Reuse Policy
-  * [GH-5](https://github.com/claranet/terraform-azurerm-aci/pull/5): Add Identity Block
+  * [GH-5](https://github.com/claranet/terraform-azurerm-aci/pull/5): Add liveness and readiness probes
+  * [GH-5](https://github.com/claranet/terraform-azurerm-aci/pull/5): Add DNS name label reuse policy
+  * [GH-5](https://github.com/claranet/terraform-azurerm-aci/pull/5): Add identity block
 
 # v7.3.0 - 2022-12-09